Complete getters and setters for rolluip transaction table types

+++ src/main/java/com/datastax/banking/dao/TransactionDao.java
@@ -218,7 +218,7 @@ public class TransactionDao {
 		a.setMin_amount(row.getDouble("min_amount"));
 		a.setMax_amount(row.getDouble("max_amount"));
 		a.setTotal_amount(row.getDouble("total_amount"));
-		a.setTotal_count(row.getDouble("total_count"));
+		a.setTotal_count(row.getLong("total_count"));
 
 		return a;
 	}
